Item interactions

+++ Assets/Scripts/Entities/HealthBar.cs
@@ -8,22 +8,24 @@ namespace Entities
 {
     public class HealthBar : MonoBehaviour
     {
+        //================================================================CLASSES
         [Serializable]
         public class HealthBarEvent : UnityEvent<Transform> { }
         
         [Serializable]
         public class HealthBarConstants
         {
-            public int maxHealth, currentHealth;
+            public float maxHealth, currentHealth;
             public bool maxAsStarting = true;
         }
-        //================================================================
+        //================================================================EDITOR VARIABLES
         [SerializeField] protected HealthBarConstants healthBarConstants;
         //TODO add material SO
         [SerializeField] protected HealthBarEvent onHealthChangedEvent;
         [SerializeField] protected HealthBarEvent onDestroyedEvent;
-        //================================================================
-        //================================================================
+        //================================================================GETTER SETTER
+        //================================================================FUNCTIONALITY
+
 
         private void Awake()
         {
@@ -33,7 +35,7 @@ namespace Entities
             }
         }
 
-        public int TakeDamage(int damage)
+        public float TakeDamage(float damage)
         {
             //TODO add material damage multiplier
             healthBarConstants.currentHealth -= damage;
